roslua can now provide a ROS service
Added and modified code to allow for providing ROS services. A test
application provides the "add two ints" service similar to the rospy
tutorial code. In fact, I have tested this with the rospy
add_two_ints_client and it works nicely. Connections are persistent
by default atm, will add a flag to control this. And then, of course,
service clients are on the list.