Optimize message deserialization
Working with actionlib and issuing hundreds of goals at a time revealed
performance problems in the deserialization process. This has been improved
by generating a format string for the whole message taking variably sized
arrays into account, minimizing calls crossing the C/Lua border. Benchmarks
show that now the slowest part is sorting the values into message instances.
Messages which have complex array fields are particularly harmful, as they
may impose many ctor calls. A solution could be to try to defer
instantiation. This would make sense if only a few fields or array fields
were to be used. Since this is not the case for us atm and the envisioned
Lua use case it should be sufficient now.

The deserialization time has been decreased by about 65% (roundtrip test
with 1000 actionlib calls as fast as possible).