lsignal: only cleanup for same Lua state
When having two Lua states in one process, the second overwrites the settings
of the first. Now, we assume to have two context only for a short time, while
the one context is initialized we keep the old one around to keep using it if
the new one fails. When the second is finalized now, the second would fail.
So we finalize only if the finalizing context is the last initializing
context.