ServiceProvider: distinguish simple and complex requests
Simple requests are request types without arrays and which use only
builtin types. They are called the way they used to, simply unpacking
the arguments.
Complex requests are request types which contain either arrays or
non-builtin types. These are called with the messages values array, as
not to destroy the structure of the message.