lua_modules: added struct.unpack_table.
Added an unpack function which returns a table of values rather than all
values. Especially in the case you have a large or unknown number of
values this is useful.

It is also required to fix a potential race condition with the Lua
garbage collector. Consder this code:
{{{
local t = {struct.unpack(format, binstring)}.
}}}
This will assign a table to t containing all read values. This seems to
be a weak table (so I didn't trace the particular piece of code where
this is decided). If the table is large, it might be that the garbage
collector has erased its values before one is done with it, causing a
segfault or glibc double linked list corruption warning.