Message: fix crash on large message deserialization
The crash was caused by a race condition between the garbage collector
and the deserialization. It seems that {unpack()} returns a weak table
which got collected before we were done with the values. This only
occurs when reading larger messages (triggered reliably at about 180kb)
and thus didn't happen until now. The solution is to use the new
struct.unpack_table() function.