ServiceClient: change return type of execute function
On successful execution the ServiceClient now by default returns the
received message. Only if the simplified_return field is set, the result
will be formatted as a table and returned unpacked. This is intended to
be used only for services with a very small return type as the previous
situation was confusing, and destroyed the structure of messages for
more complex response messages.