Subscriber: workaround for self-subscription (soft) deadlock
Let there be a roslua node which publishes a topic and subscribes to
very same topic at the same time. This causes a soft deadlock, i.e. it
tries to negotiate the connection with the "opposite" XML-RPC server,
but this request times out (hence soft) because the single-threaded
roslua node cannot process it.

For now we just avoid this situation by disallowing self-connections,
which fixes the current situation. Later we likely want a "short-cut"
version, which inter-connects our own publishers and subscribers
internally without XML-RPC calls altogether.