MsgSpec: rollback string constant changes
Turns out that string constants are not to be surrounded by quotes, they are
considered to be part of the string otherwise. Rollback recent changes to handle
this, otherwise MD5 hashes and string contents would differ from roscpp/rospy.

Extend the pattern expression for constants to cover quotes, spaces,
spaces etc.