Publisher/Subscriber: eliminate timeout deadlocks
Consider two roslua nodes. Each publishing to a topic that is subscribed
by the other node. This situation might have lead to a deadlock where
both nodes ended up waiting for data from the other node. This has been
fixed by making all connections zero timeout connections which step
through the connection process while spinning, instead of waiting for it
to complete.
This change also eliminates the workaround that was required when a node
published and subscribed to a topic at the same time (it had to be
removed from the list of publishers by the subscriber).

This is a tremendous improvement which involves lots of changes in the
low-level communication system.