SlaveProxy: allow arbitrary number of concurrent calls
Especially when connecting to a node which provides many topics allowing only
one requestTopic call at a time entails a high performance hit. Now a concurrent
call will return a request handle which encapsulates one particular call and
which can be queried on the state of the call. No spinning needs to be called
from the outside, this is done automatically.

Updated the Subscriber to make use of the new SlaveProxy.