Message: fix deserialization format string split bug
The format string splitting already took an error case into account, i.e. if a
"c0" in the following format string would not be prefixed by a number. Only if
this was not the case, it checked if there was an incomplete number format, e.g.
"i" instead of "i4". This ordering can cause problems. Consider a string where
splitting would occur at "i|4c0". Now, the c0 would not be recognized, because
overhang is "4c", afterwards, the 4 is taken over because otherwise "i" is
incomplete, again leaving an unprefixed "c0" in the following string.

The solution is to first check for incomplete types, and then check for "c0"
overhang, then we would get "i|4c0", extend by 4 to "i4|c0", extend by c0 to
"i4c0|".

The problem was hard to find because the error was caught and not exposed.
Changed this so that if deserialization fails we get a trace and make it obvious
by printing an error. Need to verify that it does not print too many false
errors in further tests.