TcpRosPubSubConnection: fix error if remote closes connection
TcpRosPubSubConnection called TcpRosConnection.receive in a wrong manner,
expecting errors to be thrown rather than checking the return value of the
methot itself. This could cause a nil buffer to be passed to message
deserialization. This bug was uncovered with the changes of the previous commit
exposing errors rather than silently ignoring them.