ServiceClient: better handling of died connections
Consider the following: a persistent service client calls a provider.
In between two calls, the provider is restarted, hence rendering
existing connections closed. Before this commit, this would be reported
as a service execution failure. Rather, we want to make the
ServiceClient more reliable and try establishing a new connection and
report failure only if this fails.