utils: C package loader looks for *_lua.luaso
To import a module abc, the module can now also be in a package abc or
abc_lua and named lib/abc.luaso or lib/abc_lua.luaso. This is done to
allow for situations were you have a library for C++ and Lua, and the
Lua module uses the C++ module and both have the same basename. Now
CMake tries to be clever and optimizes the "-labc" out in the linking
command of the Lua library.