message: print byte[] and uint8[] as number of bytes
These data types are typically used to transfer large amounts of binary
data, which is irrelevant to print and can print so much (e.g. for
messages containing point clouds) that performance takes a bad hit and
the output becomes cluttered and virtually useless.

To remedy this, just print "N bytes of data" instead.